requirements.txt
Flask
python-dotenv
requests
PyMySQL
python-magic